refactor: moved gson kotlin fix to extensions
